rename: NOTICE_MESSAGES 파일 이동
